Move stdout/err -> logfile redirect context manager to core.py
